Move OutputDialog to ui package
